% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrrow.R
\name{marrow_dir}
\alias{marrow_dir}
\title{Iteratively collate output of function into an Arrow dataset out of memory}
\usage{
marrow_dir(.x, .f, ..., .path, .partitioning = c(), .format = "parquet")
}
\arguments{
\item{.x}{vector or list of values for .f to iterate over}

\item{.f}{function; must return a data.frame/tibble}

\item{...}{other arguments to .f}

\item{.path}{path to directory where collated Arrow dataset will be stored.
will be created if it does not exist}

\item{.partitioning}{character vector of columns to use for partitioning.
Columns must exist in output of .f.}

\item{.format}{"parquet" (the default) or "arrow".}
}
\value{
path to new dataset directory; character string of length one.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
map + arrow: iterate over a function and collate the results into
an Arrow dataset. This happens without the whole dataset being in memory,
so is suitable for large data objects. The function must return a data.frame or
tibble. The returned value is a path to the directory containing the
Arrow dataset.
}
\examples{
# ADD_EXAMPLES_HERE
}
