% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pin.arrow_dplyr_query}
\alias{pin.arrow_dplyr_query}
\title{FUNCTION_TITLE}
\usage{
\method{pin}{arrow_dplyr_query}(x, name = NULL, description = NULL, board = NULL, ...)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{name}{DESCRIPTION.}

\item{description}{DESCRIPTION.}

\item{board}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
