% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pin_preview.arrow_dataset}
\alias{pin_preview.arrow_dataset}
\title{FUNCTION_TITLE}
\usage{
\method{pin_preview}{arrow_dataset}(x, ...)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
