% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrrow-package.R
\docType{package}
\name{purrrow-package}
\alias{purrrow}
\alias{purrrow-package}
\title{purrrow: Out-of-memory data collation into Arrow datasets}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://petrbouchal.xyz/purrrow}
  \item \url{https://github.com/petrbouchal/purrrow}
  \item Report bugs at \url{https://github.com/petrbouchal/purrrow/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
