% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrrow-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{board_initialize}
\alias{board_pin_create}
\alias{board_pin_get}
\alias{board_pin_find}
\alias{board_pin_remove}
\alias{board_pin_versions}
\alias{pin_get}
\alias{pin}
\alias{pin_load}
\alias{pin_preview}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pins}{\code{\link[pins:custom-boards]{board_initialize}}, \code{\link[pins:custom-boards]{board_pin_create}}, \code{\link[pins:custom-boards]{board_pin_find}}, \code{\link[pins:custom-boards]{board_pin_get}}, \code{\link[pins:custom-boards]{board_pin_remove}}, \code{\link[pins:custom-boards]{board_pin_versions}}, \code{\link[pins]{pin}}, \code{\link[pins]{pin_get}}, \code{\link[pins:custom-pins]{pin_load}}, \code{\link[pins:custom-pins]{pin_preview}}}
}}

